master

+++ modules/rbac/controllers/RoleController.php
@@ -92,15 +92,16 @@ class RoleController extends \backend\components\Controller
         $model = new AuthItem();
         $model->type = 1;
 
-        $post = Yii::$app->request->post();
-        if (
-            $model->load($post)
-            && $model->save()
-        ) {
-            $model->deleteAllChildren();
-            $model->addChildren($post['child-roles-names']);
-            $model->addChildren($post['child-permissions-names']);
-            return $this->redirect(['view', 'id' => $model->id]);
+        if (Yii::$app->request->isPost) {
+            $post = Yii::$app->request->post();
+            if ($model->load($post) && $model->save()) {
+                $model->deleteAllChildren();
+                $model->addChildren($post['child-roles-names']);
+                $model->addChildren($post['child-permissions-names']);
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
+        } else {
+            $model->loadDefaultValues();
         }
 
         $rolesNotOfThisRole = $permissionsNotOfThisRole = [];
